Added docu Cleanups More unicode Started fixing this big unod/redo bug Some renamings Indentation of source code
svn path=/trunk/koffice/; revision=45182

+++ kspread/kspread_dlg_create.cc
@@ -976,7 +976,7 @@ QString KSpreadcreate::make_formula( const QString& _text,int nb_line)
 		//so don't find "
 		text+="\"";
 	    }
-	    else if(pos_car < text.length())
+	    else if(pos_car < (int)text.length())
 	    {
 		//todo
 	    }
